Revert "fix: 2 - DEREFERENCE_NULLABLE - long value = longAdapter.read(in).longValue(); - dereferenced expression longAdapter.read(in) is @Nullable"
This reverts commit dd527b4851f5a361b855eda2b147829888deaa5b.

+++ gson/src/main/java/com/google/gson/Gson.java
@@ -453,7 +453,7 @@ public final class Gson {
         List<Long> list = new ArrayList<Long>();
         in.beginArray();
         while (in.hasNext()) {
-          Long value = longAdapter.read(in);
+          long value = longAdapter.read(in).longValue();
           list.add(value);
         }
         in.endArray();
